rename 'okta-auth' plugin to 'auth'
