+++ server/plugins/auth/get-stubbed-credentials.js
