+++ server/plugins/auth/get-token-from-header.js
