+++ server/plugins/auth/verify-scopes.js
