+++ test/unit/plugins/auth-okta/okta-jwt-verifier.test.js
@@ -1,6 +1,6 @@
 describe('Okta JWT Verifier', () => {
   test('instantiates verifier without error', () => {
-    const oktaJwtVerifier = require('../../../../server/plugins/auth-okta/okta-jwt-verifier')
+    const oktaJwtVerifier = require('../../../../server/plugins/auth/okta-jwt-verifier')
     expect(oktaJwtVerifier).toBeDefined()
   })
 })
