+++ test/unit/plugins/auth-okta/verify-audience.test.js
@@ -1,5 +1,5 @@
 
-const verifyAudience = require('../../../../server/plugins/auth-okta/verify-audience')
+const verifyAudience = require('../../../../server/plugins/auth/verify-audience')
 describe('Verify scopes', () => {
   test('throws no error if audiences match', () => {
     verifyAudience('aud1', 'aud1')
