+++ test/unit/plugins/auth-okta/verify-scopes.test.js
@@ -1,5 +1,5 @@
 
-const verifyScopes = require('../../../../server/plugins/auth-okta/verify-scopes')
+const verifyScopes = require('../../../../server/plugins/auth/verify-scopes')
 describe('Verify scopes', () => {
   test('throws no error if all scopes present', () => {
     verifyScopes(['scopeOne', 'scopeTwo'], ['scopeTwo', 'scopeOne'])
