+++ test/unit/server/routes/schedule.test.js
@@ -4,8 +4,8 @@ describe('Schedule test', () => {
   let scheduleService
 
   beforeAll(async () => {
-    const oktaJwtVerifier = require('../../../../server/plugins/auth-okta/okta-jwt-verifier')
-    jest.mock('../../../../server/plugins/auth-okta/okta-jwt-verifier')
+    const oktaJwtVerifier = require('../../../../server/plugins/auth/okta-jwt-verifier')
+    jest.mock('../../../../server/plugins/auth/okta-jwt-verifier')
     oktaJwtVerifier.verifyAccessToken.mockImplementation(() =>
       Promise.resolve({ claims: { roles: ['payment-admin'] } })
     )
